+++ src/components/TrackerCard/TrackerCardContent.tsx
@@ -36,15 +36,41 @@ const TrackerCardContent: FC<Props> = ({
     }
   }
 
+  const remainingCompletions = requiredCompletions
+    .map((rc) => {
+      const todayCompletion = aggTodayCompletions.find((c) => rc.unit === c.unit);
+      if (todayCompletion) {
+        const remain = rc.quantity - todayCompletion.quantity;
+        return {
+          ...rc,
+          quantity: remain
+        } as Completion;
+      }
+      return rc;
+    })
+    .filter((c) => c.quantity > 0); // don't take completed completions
+
   return (
     <CardContent {...cardContentProps}>
-      <Typography>Requis :</Typography>
-      <CompletionChipList
-        completions={requiredCompletions}
-        onChipClick={onChipClick}
-        selectedCompletions={selectedCompletions}
-      />
-      {todayCompletions.length > 0 && (
+      {requiredCompletions.length > 0 && (
+        <>
+          <Typography>Requis :</Typography>
+          <CompletionChipList
+            completions={requiredCompletions}
+            onChipClick={onChipClick}
+            selectedCompletions={selectedCompletions}
+          />
+        </>
+      )}
+      {remainingCompletions.length > 0 ? (
+        <>
+          <Typography>Restants :</Typography>
+          <CompletionChipList
+            completions={remainingCompletions}
+            requiredCompletions={requiredCompletions}
+          />
+        </>
+      ) : (
         <>
           <Typography>Effectués :</Typography>
           <CompletionChipList completions={aggTodayCompletions} />