+++ src/components/forms/TrackerForm/TrackerForm.tsx
@@ -1,12 +1,14 @@
-import DatePicker from '@mui/lab/DatePicker';
 import { Box, Button, Stack, TextField } from '@mui/material';
+import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { FC } from 'react';
-import { Controller, useFieldArray, useForm } from 'react-hook-form';
+import { Controller, useFieldArray, useForm, useWatch } from 'react-hook-form';
 import { v4 } from 'uuid';
 
 import { useAppDispatch } from '../../../app/hooks';
+import Completion from '../../../models/Completion';
 import TrackerStatus from '../../../models/TrackerStatus';
 import { createTracker } from '../../../store/trackers/trackersSlice';
+import DefaultCompletionsForm from '../DefaultCompletionsForm/DefaultCompletionsForm';
 import RequiredCompletionsForm from '../RequiredCompletionsForm/RequiredCompletionsForm';
 import { FormValues } from './types';
 
@@ -14,14 +16,10 @@ const getDefaultValues = (): FormValues => ({
   id: v4(),
   beginDate: new Date().toString(),
   duration: '',
+  defaultCompletions: [],
   entries: [],
   name: '',
-  requiredCompletions: [
-    {
-      quantity: '1',
-      unit: 'fois'
-    }
-  ],
+  requiredCompletions: [],
   status: TrackerStatus.active
 });
 
@@ -30,18 +28,33 @@ interface Props {
 }
 
 const TrackerForm: FC<Props> = ({ hideForm }) => {
+  const dispatch = useAppDispatch();
   const { control, handleSubmit, reset } = useForm<FormValues>({
     defaultValues: getDefaultValues()
   });
-  const { fields, append, remove } = useFieldArray({
+  const requiredCompletionsFieldArray = useFieldArray({
     control, // control props comes from useForm
     name: 'requiredCompletions'
   });
-
-  const dispatch = useAppDispatch();
+  const defaultCompletionsFieldArray = useFieldArray({
+    control, // control props comes from useForm
+    name: 'defaultCompletions'
+  });
+  const requiredCompletions = useWatch({
+    control,
+    name: 'requiredCompletions'
+  });
 
   const resetToDefault = () => {
     reset(getDefaultValues());
+  };
+
+  const removeRequiredCompletion = (index?: number | number[]) => {
+    // If it's the latest requiredCompletions, defaultCompletions should be emptied
+    if (requiredCompletions.length === 1) {
+      defaultCompletionsFieldArray.replace([]);
+    }
+    requiredCompletionsFieldArray.remove(index);
   };
 
   const onSubmit = (data: FormValues) => {
@@ -139,7 +152,36 @@ const TrackerForm: FC<Props> = ({ hideForm }) => {
         }}
       />
 
-      <RequiredCompletionsForm append={append} control={control} fields={fields} remove={remove} />
+      <RequiredCompletionsForm
+        append={requiredCompletionsFieldArray.append}
+        control={control}
+        fields={requiredCompletionsFieldArray.fields}
+        remove={removeRequiredCompletion}
+        requiredCompletions={requiredCompletions.map(
+          (rc) =>
+            ({
+              ...rc,
+              quantity: Number(rc.quantity)
+            } as Completion)
+        )}
+      />
+
+      {requiredCompletions.length > 0 &&
+        requiredCompletions.some((rc) => rc.unit !== '' && rc.unit !== undefined) && (
+          <DefaultCompletionsForm
+            append={defaultCompletionsFieldArray.append}
+            control={control}
+            fields={defaultCompletionsFieldArray.fields}
+            remove={defaultCompletionsFieldArray.remove}
+            requiredCompletions={requiredCompletions.map(
+              (rc) =>
+                ({
+                  ...rc,
+                  quantity: Number(rc.quantity)
+                } as Completion)
+            )}
+          />
+        )}
 
       <Stack direction="row" justifyContent="center" spacing={1}>
         <Button type="submit" onClick={handleSubmit(onSubmit)} variant={'outlined'}>