+++ src/config/Constants.ts
@@ -5,6 +5,8 @@ export const JSON_DATA_EXTENSION = '.bj.json'; // Bujo Tracker json
 export const PAYPAL_DONATE_LINK =
   'https://www.paypal.com/donate/?business=UF2GH7ZPMLES8&no_recurring=0&item_name=Chaque+don+me+permet+de+continuer+%C3%A0+d%C3%A9velopper+et+%C3%A0+am%C3%A9liorer+l%27application+Bujo+Tracker%2C+merci+%21+%3A%29&currency_code=EUR';
 
+export const DEFAULT_COMPLETION_NAME = 'fois';
+
 // ===== Dates
 export const SEVEN_DAYS_AGO_DATE: Date = subDays(new Date(), 7);
 export const SEVEN_DAYS_AGO_STRING: string = SEVEN_DAYS_AGO_DATE.toString();